+++ include/pyawaitable/genwrapper.h
@@ -1,0 +1,10 @@
+#ifndef PYAWAITABLE_GENWRAPPER_H
+#define PYAWAITABLE_GENWRAPPER_H
+
+#include <Python.h>
+
+extern PyTypeObject _AwaitableGenWrapperType;
+
+PyObject *gen_next(PyObject *self);
+
+#endif
