+++ src/_pyawaitable/awaitable.c
@@ -1,0 +1,181 @@
+/* *INDENT-OFF* */
+#include <Python.h>
+#include <pyawaitable/backport.h>
+#include <pyawaitable/awaitableobject.h>
+
+PyDoc_STRVAR(awaitable_doc,
+    "Awaitable transport utility for the C API.");
+
+static PyObject *
+awaitable_new_func(PyTypeObject *tp, PyObject *args, PyObject *kwds)
+{
+    assert(tp != NULL);
+    assert(tp->tp_alloc != NULL);
+
+    PyObject *self = tp->tp_alloc(tp, 0);
+    if (self == NULL) {
+        return NULL;
+    }
+
+    AwaitableObject *aw = (AwaitableObject *) self;
+    aw->aw_callbacks = NULL;
+    aw->aw_callback_size = 0;
+    aw->aw_result = NULL;
+    aw->aw_gen = NULL;
+    aw->aw_values = NULL;
+    aw->aw_values_size = 0;
+    aw->aw_state = 0;
+    aw->aw_done = false;
+
+    return (PyObject *) aw;
+}
+
+static PyObject *
+awaitable_next(PyObject *self)
+{
+    AwaitableObject *aw = (AwaitableObject *) self;
+
+
+    if (aw->aw_done) {
+        PyErr_SetString(PyExc_RuntimeError, "cannot reuse awaitable");
+        return NULL;
+    }
+
+    PyObject* gen = _awaitable_genwrapper_new(aw);
+
+    if (gen == NULL) {
+        return NULL;
+    }
+
+    aw->aw_gen = Py_NewRef(gen);
+    aw->aw_done = true;
+    return gen;
+}
+
+static void
+awaitable_dealloc(PyObject *self)
+{
+    AwaitableObject *aw = (AwaitableObject *) self;
+    if (aw->aw_values) {
+        for (int i = 0; i < aw->aw_values_size; i++)
+            Py_DECREF(aw->aw_values[i]);
+        PyMem_Free(aw->aw_values);
+    }
+
+    Py_XDECREF(aw->aw_gen);
+    Py_XDECREF(aw->aw_result);
+
+    for (int i = 0; i < aw->aw_callback_size; i++) {
+        awaitable_callback *cb = aw->aw_callbacks[i];
+        if (!cb->done) Py_DECREF(cb->coro);
+        PyMem_Free(cb);
+    }
+
+    if (aw->aw_arb_values) PyMem_Free(aw->aw_arb_values);
+    Py_TYPE(self)->tp_free(self);
+}
+
+void
+_awaitable_cancel(PyObject *aw)
+{
+    assert(aw != NULL);
+    Py_INCREF(aw);
+
+    AwaitableObject *a = (AwaitableObject *) aw;
+
+    for (int i = 0; i < a->aw_callback_size; i++) {
+        awaitable_callback* cb = a->aw_callbacks[i];
+        if (!cb->done)
+            Py_DECREF(cb->coro);
+    }
+
+    PyMem_Free(a->aw_callbacks);
+    a->aw_callback_size = 0;
+    Py_DECREF(aw);
+}
+
+
+PyTypeObject _AwaitableType = {
+    PyVarObject_HEAD_INIT(NULL, 0)
+    .tp_name = "_awaitable",
+    .tp_basicsize = sizeof(AwaitableObject),
+    .tp_dealloc = awaitable_dealloc,
+    .tp_as_async = &async_methods,
+    .tp_flags = Py_TPFLAGS_DEFAULT,
+    .tp_doc = awaitable_doc,
+    .tp_iternext = awaitable_next,
+    .tp_new = awaitable_new_func,
+    .tp_methods = awaitable_methods
+};
+
+
+int
+_awaitable_await(
+    PyObject *aw,
+    PyObject *coro,
+    awaitcallback cb,
+    awaitcallback_err err
+)
+{
+    assert(aw != NULL);
+    assert(coro != NULL);
+    Py_INCREF(coro);
+    Py_INCREF(aw);
+    AwaitableObject *a = (AwaitableObject *) aw;
+
+    awaitable_callback *aw_c = PyMem_Malloc(sizeof(awaitable_callback));
+    if (aw_c == NULL) {
+        Py_DECREF(aw);
+        Py_DECREF(coro);
+        PyErr_NoMemory();
+        return -1;
+    }
+
+    ++a->aw_callback_size;
+    if (a->aw_callbacks == NULL) {
+        a->aw_callbacks = PyMem_Calloc(a->aw_callback_size,
+        sizeof(awaitable_callback *));
+    } else {
+        a->aw_callbacks = PyMem_Realloc(a->aw_callbacks,
+        sizeof(awaitable_callback *) * a->aw_callback_size
+    );
+    }
+
+    if (a->aw_callbacks == NULL) {
+        --a->aw_callback_size;
+        Py_DECREF(aw);
+        Py_DECREF(coro);
+        PyMem_Free(aw_c);
+        PyErr_NoMemory();
+        return -1;
+    }
+
+    aw_c->coro = coro; // steal our own reference
+    aw_c->callback = cb;
+    aw_c->err_callback = err;
+    a->aw_callbacks[a->aw_callback_size - 1] = aw_c;
+    Py_DECREF(aw);
+
+    return 0;
+}
+
+int
+_awaitable_set_result(PyObject *awaitable, PyObject *result)
+{
+    assert(awaitable != NULL);
+    assert(result != NULL);
+    Py_INCREF(result);
+    Py_INCREF(awaitable);
+
+    AwaitableObject *aw = (AwaitableObject *) awaitable;
+    if (aw->aw_gen == NULL) {
+        PyErr_SetString(PyExc_TypeError, "no generator is currently present");
+        Py_DECREF(awaitable);
+        Py_DECREF(result);
+        return -1;
+    }
+    _awaitable_genwrapper_set_result(aw->aw_gen, result);
+    Py_DECREF(awaitable);
+    Py_DECREF(result);
+    return 0;
+}
